Refactor Today pure logic helpers
